+++ src/PolyClone.cpp
@@ -2,8 +2,10 @@
 // Licensed under GNU GPLv3
 
 #include "plugin.hpp"
+#include "CloneModule.hpp"
 
-struct PolyClone : VenomModule {
+struct PolyClone : CloneModuleBase {
+
   enum ParamId {
     CLONE_PARAM,
     PARAMS_LEN
@@ -55,16 +57,19 @@ struct PolyClone : VenomModule {
       for (int j=0; j<clones; j++)
         outputs[POLY_OUTPUT].setVoltage(val, c++);
     }
-    outputs[POLY_OUTPUT].setChannels(goodCh * clones);
+    int outCnt = goodCh * clones;
+    outputs[POLY_OUTPUT].setChannels(outCnt);
+    processExpander(clones, goodCh);
 
     if (lightDivider.process()) {
       for (int i=1; i<16; i++) {
         lights[CHANNEL_LIGHTS+i*2].setBrightness(i<goodCh);
         lights[CHANNEL_LIGHTS+i*2+1].setBrightness(i>=goodCh && i<ch);
       }
+      setExpanderLights(goodCh);
     }
   }
-
+  
 };